Rename junit5 to junit-jupiter
