+++ fixture-monkey-junit5/gradle.properties
@@ -1,3 +1,3 @@
-artifactId=fixture-monkey-junit5
-artifactName=Fixture Monkey JUnit5
-artifactDescription=Fixture Monkey extends Junit5 to support junit5 api.
+artifactId=fixture-monkey-junit-jupiter
+artifactName=Fixture Monkey JUnit Jupiter
+artifactDescription=Fixture Monkey supports junit-jupiter-api
