+++ fixture-monkey-junit5/src/test/java/com/navercorp/fixturemonkey/junit/jupiter/extension/FixtureMonkeyParameterExtensionTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.navercorp.fixturemonkey.junit5.extension;
+package com.navercorp.fixturemonkey.junit.jupiter.extension;
 
 import static org.assertj.core.api.BDDAssertions.then;
 
@@ -36,7 +36,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import lombok.Data;
 
-import com.navercorp.fixturemonkey.junit5.annotation.GiveMe;
+import com.navercorp.fixturemonkey.junit.jupiter.annotation.GiveMe;
 
 @ExtendWith(FixtureMonkeyParameterExtension.class)
 class FixtureMonkeyParameterExtensionTest {
