+++ settings.gradle
@@ -9,4 +9,4 @@ include "fixture-monkey-jackson"
 include "fixture-monkey-autoparams"
 include "fixture-monkey-mockito"
 include "fixture-monkey-starter"
-include "fixture-monkey-junit5"
+include "fixture-monkey-junit-jupiter"
